Id type changed to long.

+++ src/main/java/server/filestorm/model/repository/UserRepository.java
@@ -8,13 +8,13 @@ import org.springframework.data.jpa.repository.Query;
 
 import server.filestorm.model.entity.User;
 
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
     
     @Query("SELECT u FROM User u WHERE u.username = ?1")
     Optional<User> findUserByUsername(String username);
 
     @Query("SELECT u FROM User u WHERE u.username = ?1 AND u.id = ?2")
-    Optional<User> findUserByUsernameAndId(String username, Integer id);
+    Optional<User> findUserByUsernameAndId(String username, Long id);
     
     @Query("SELECT u.username, u.id FROM User u WHERE u.username ILIKE CONCAT('%', ?1, '%') ORDER BY u.username ASC")
     Optional<List<Object[]>> queryUsersByUsername(String username);
